// src/pages/Details/tabs/Packaging/types.ts
export const DEFAULT_AB_RESULTS: ABTestResults = { titles: [], thumbnails: [] };
// ...
export interface ABTestResults {
    titles: number[];
    thumbnails: number[];
}